Update README.md
The question:
a) Program should be able to assign seats on each flight of the airline's only plane (capacity 20 seats). 

b) If the person types 1, then the program should assign a seat in the first class section (seats 1-5). If the person types 2, then the program should assign a seat the economy section (seats 6-20). The program should then print a boarding pass including the person's seat number and whether it is in the first class or economy section of the plane. 

c) Using a single-subscripted array to represent the seating chart of the plane. Initializing all the elements of the array to 0 to indicate that all seats are empty. As each seat is assigned, set the corresponding elements of the array to 1 to indicate that the seat is no longer available. 

d) The program should, of course, never assign a seat that has already been assigned. When the first class section is full, the program should ask the person if it is acceptable to be placed in the economy section (vice versa). If yes, then make the appropriate seat assignment. If no, then print the message "Next flight leaves in 3 hours."